unpacker-worker/requirements.txt
pynsq
frida
python-snappy